Final version with bonus
Put a react app in another folder "simple calc" and it caused errors.
