+++ README.md
@@ -1,87 +1,70 @@
-# ![GA Logo](https://ga-dash.s3.amazonaws.com/production/assets/logo-9f88ae6c9c3871690e33280fcf557f33.png) Simple Calc
+# Getting Started with Create React App
 
-Now, it's time for you to check back on everything! You will be building a calculator with React, and only minimal instructions have been provided for you to really think about what's happening.
+This project was bootstrapped with [Create React App](https://github.com/facebook/create-react-app).
 
-At first, your calculator will just add 2 numbers together when they are typed in. For the bonus, you might decide to get more creative.
+## Available Scripts
 
-## Set Up
+In the project directory, you can run:
 
-Fork & Clone this repo, cd into repo, run `npx create-react-app .` to turn your cloned repo into a react app.
+### `npm start`
 
-## Step 1
+Runs the app in the development mode.\
+Open [http://localhost:3000](http://localhost:3000) to view it in your browser.
 
-Start by creating a single component file in the `src` directory, and name it `Calculator.js`. In this file, create your `Calculator` class. Use the `App.js` as an example of how to create a basic component. Add the following JSX to your Calculator's `render()` function:
+The page will reload when you make changes.\
+You may also see any lint errors in the console.
 
-```jsx
-<div className="container">
-  <h1>Add with React!</h1>
+### `npm test`
 
-  <div className="add">
-    <input type="text" />
-    <span>+</span>
-    <input type="text" />
-    <button>=</button>
-    <h3>Addition results go here!</h3>
-  </div>
-</div>
-```
+Launches the test runner in the interactive watch mode.\
+See the section about [running tests](https://facebook.github.io/create-react-app/docs/running-tests) for more information.
 
-## Step 2
+### `npm run build`
 
-Set up the initial state of your component. What state attributes will you need to track? What values should those state items start with? Where is that state displayed in the browser?
+Builds the app for production to the `build` folder.\
+It correctly bundles React in production mode and optimizes the build for the best performance.
 
-> Hint: You will only need one state. That one state can hold as many key-value pairs as you need!
+The build is minified and the filenames include the hashes.\
+Your app is ready to be deployed!
 
-## Step 3
+See the section about [deployment](https://facebook.github.io/create-react-app/docs/deployment) for more information.
 
-You will want to trigger a function when the values in your textboxes change. You can capture these values by setting a function on the onChange property. Let's say I have a textbox tracking my first number.
+### `npm run eject`
 
-```jsx
-<input type="number"
-  name="num1"
-  placeholder="Enter your first number"
-  value={this.state.num1}
-  onChange={ (e) => this.setNum(e, 'num1') }
-/>
-```
+**Note: this is a one-way operation. Once you `eject`, you can't go back!**
 
-I want to store this number as part of my state. Let's say I decided to call it `num1`. I could set my state like so:
+If you aren't satisfied with the build tool and configuration choices, you can `eject` at any time. This command will remove the single build dependency from your project.
 
-```text
-setNum = (e, num) => {
-  this.setState({ [num]: e.target.value});
-}
-```
+Instead, it will copy all the configuration files and the transitive dependencies (webpack, Babel, ESLint, etc) right into your project so you have full control over them. All of the commands except `eject` will still work, but they will point to the copied scripts so you can tweak them. At this point you're on your own.
 
-> Hint: The \[\] are there so we can use a dynamic key value! This value becomes `num1` or `num2` depending on what was clicked and sent to the function from `onChange`.
+You don't have to ever use `eject`. The curated feature set is suitable for small and middle deployments, and you shouldn't feel obligated to use this feature. However we understand that this tool wouldn't be useful if you couldn't customize it when you are ready for it.
 
-The above method relies on a [closure](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Closures) to determine which button is being clicked, which is a common technique in React. Can you think of another way(s) to accomplish this task without using a closure?
+## Learn More
 
-If you decided to use buttons for your calculator, you probably want to use `onClick` instead of `onChange`, but the concepts are the same! Here is some documentation to help you choose what you want to do and how to do it:
+You can learn more in the [Create React App documentation](https://facebook.github.io/create-react-app/docs/getting-started).
 
-* [React form documentation](https://facebook.github.io/react/docs/forms.html)
-* [A list of events React supports](https://facebook.github.io/react/docs/events.html#supported-events)
+To learn React, check out the [React documentation](https://reactjs.org/).
 
-## Step 4
+### Code Splitting
 
-Once you've got your event handlers set up to capture the input, you'll need to create a method for your submit button. The method should accept the triggered event, get the input values from your state, add them together, and set part of the state to the new `sum`.
+This section has moved here: [https://facebook.github.io/create-react-app/docs/code-splitting](https://facebook.github.io/create-react-app/docs/code-splitting)
 
-Hint: Where should this method go?
+### Analyzing the Bundle Size
 
-In the same component as it's being used - between the constructor and the render.
+This section has moved here: [https://facebook.github.io/create-react-app/docs/analyzing-the-bundle-size](https://facebook.github.io/create-react-app/docs/analyzing-the-bundle-size)
 
-> Thought: How will you handle inputs that aren't numbers?
+### Making a Progressive Web App
 
-## Step 5
+This section has moved here: [https://facebook.github.io/create-react-app/docs/making-a-progressive-web-app](https://facebook.github.io/create-react-app/docs/making-a-progressive-web-app)
 
-Once the state of the `sum` has been set, React will re-render the whole component. Make sure you have a place in your JSX that displays the result!
+### Advanced Configuration
 
-## Bonus
+This section has moved here: [https://facebook.github.io/create-react-app/docs/advanced-configuration](https://facebook.github.io/create-react-app/docs/advanced-configuration)
 
-* Make the calculator work with any of the 4 basic arithmetic operations
+### Deployment
 
-  \(+, -, \*, /\). How will this change your `state` and your JSX?
+This section has moved here: [https://facebook.github.io/create-react-app/docs/deployment](https://facebook.github.io/create-react-app/docs/deployment)
 
-## Licensing
-1. All content is licensed under a CC-BY-NC-SA 4.0 license.
-2. All software code is licensed under GNU GPLv3. For commercial use or alternative licensing, please contact legal@ga.co.
+### `npm run build` fails to minify
+
+This section has moved here: [https://facebook.github.io/create-react-app/docs/troubleshooting#npm-run-build-fails-to-minify](https://facebook.github.io/create-react-app/docs/troubleshooting#npm-run-build-fails-to-minify)
