+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "simple-calc",
+  "name": "react-simple-calc-lab",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
